# packages/api-client/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SchulSaniApp API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  owner: 'owner',
  admin: 'admin',
  sanitaeter_leitung_admin: 'sanitaeter_leitung_admin',
  sanitaeter_leitung: 'sanitaeter_leitung',
  teacher: 'teacher',
  sanitaeter: 'sanitaeter',
} as const;

export interface CustomRole {
  id: string;
  name: string;
  color: string;
  permissions: string[];
  createdBy: string;
  createdAt: string;
}

export interface User {
  id: string;
  iservUsername: string;
  displayName: string;
  email: string;
  role: UserRole;
  customRoleId?: string | null;
  customRole?: CustomRole | null;
  isActive: boolean;
  createdAt: string;
}

export interface LoginResponse {
  token: string;
  user: User;
}

export interface MissionResponder {
  id: string;
  userId: string;
  displayName: string;
  respondedAt: string;
}

export type MissionStatus = typeof MissionStatus[keyof typeof MissionStatus];


export const MissionStatus = {
  active: 'active',
  closed: 'closed',
} as const;

export interface Mission {
  id: string;
  title: string;
  location: string;
  patientInitials: string;
  treatmentNotes: string;
  status: MissionStatus;
  createdBy: string;
  createdByName: string;
  createdAt: string;
  closedAt?: string | null;
  responders: MissionResponder[];
}

export interface CreateMissionRequest {
  title: string;
  location: string;
  patientInitials: string;
  treatmentNotes?: string;
}

export interface UpdateMissionRequest {
  title?: string;
  location?: string;
  patientInitials?: string;
  treatmentNotes?: string;
}

export interface DutySlot {
  id: string;
  userId: string;
  displayName: string;
  date: string;
  startTime: string;
  endTime: string;
  notes?: string | null;
  createdAt: string;
}

export interface CreateDutySlotRequest {
  userId: string;
  date: string;
  startTime: string;
  endTime: string;
  notes?: string;
}

export interface DutyEntry {
  id: string;
  userId: string;
  dutySlotId: string;
  date: string;
  enteredAt: string;
  confirmedAt?: string | null;
}

export interface DutyTodayResponse {
  slot: DutySlot | null;
  entry: DutyEntry | null;
}

export interface DutyEntryWithUser {
  id: string;
  userId: string;
  displayName?: string | null;
  dutySlotId: string;
  date: string;
  enteredAt: string;
  confirmedAt?: string | null;
}

export type ActivityLogEntryMetadata = { [key: string]: unknown } | null;

export interface ActivityLogEntry {
  id: string;
  userId?: string | null;
  displayName?: string | null;
  action: string;
  metadata?: ActivityLogEntryMetadata;
  createdAt: string;
}

export type LoaRequestStatus = typeof LoaRequestStatus[keyof typeof LoaRequestStatus];


export const LoaRequestStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface LoaRequest {
  id: string;
  userId: string;
  displayName: string;
  reason: string;
  fromDate: string;
  toDate: string;
  status: LoaRequestStatus;
  reviewedBy?: string | null;
  reviewedByName?: string | null;
  reviewedAt?: string | null;
  createdAt: string;
}

export interface CreateLoaRequest {
  reason: string;
  fromDate: string;
  toDate: string;
}

export type UpdateLoaRequestStatus = typeof UpdateLoaRequestStatus[keyof typeof UpdateLoaRequestStatus];


export const UpdateLoaRequestStatus = {
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface UpdateLoaRequest {
  status: UpdateLoaRequestStatus;
}

export interface NewsPost {
  id: string;
  authorId: string;
  authorName: string;
  title: string;
  content: string;
  isPinned: boolean;
  createdAt: string;
}

export interface CreateNewsRequest {
  title: string;
  content: string;
  isPinned?: boolean;
}

export interface UpdateNewsRequest {
  title?: string;
  content?: string;
  isPinned?: boolean;
}

export type UpdateUserRequestRole = typeof UpdateUserRequestRole[keyof typeof UpdateUserRequestRole];


export const UpdateUserRequestRole = {
  admin: 'admin',
  sanitaeter_leitung_admin: 'sanitaeter_leitung_admin',
  sanitaeter_leitung: 'sanitaeter_leitung',
  teacher: 'teacher',
  sanitaeter: 'sanitaeter',
} as const;

export interface UpdateUserRequest {
  role?: UpdateUserRequestRole;
  isActive?: boolean;
  customRoleId?: string | null;
}

export interface CreateCustomRoleRequest {
  name: string;
  color?: string;
  permissions: string[];
}

export interface UpdateCustomRoleRequest {
  name?: string;
  color?: string;
  permissions?: string[];
}

export type PushTokenRequestPlatform = typeof PushTokenRequestPlatform[keyof typeof PushTokenRequestPlatform];


export const PushTokenRequestPlatform = {
  ios: 'ios',
  android: 'android',
} as const;

export interface PushTokenRequest {
  token: string;
  platform: PushTokenRequestPlatform;
}

export interface MessageResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export type GetActivityParams = {
userId?: string;
action?: string;
};