# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SchulSaniApp API
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    CustomRole:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
        permissions:
          type: array
          items:
            type: string
        createdBy:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - color
        - permissions
        - createdBy
        - createdAt
    User:
      type: object
      properties:
        id:
          type: string
        iservUsername:
          type: string
        displayName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [owner, admin, sanitaeter_leitung_admin, sanitaeter_leitung, teacher, sanitaeter]
        customRoleId:
          type: string
          nullable: true
        customRole:
          $ref: "#/components/schemas/CustomRole"
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - iservUsername
        - displayName
        - email
        - role
        - isActive
        - createdAt
...
    DutySlot:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        displayName:
          type: string
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - userId
        - displayName
        - date
        - startTime
        - endTime
        - createdAt
    CreateDutySlotRequest:
      type: object
      properties:
        userId:
          type: string
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        notes:
          type: string
      required:
        - userId
        - date
        - startTime
        - endTime
    DutyEntry:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        dutySlotId:
          type: string
        date:
          type: string
        enteredAt:
          type: string
        confirmedAt:
          type: string
          nullable: true
      required:
        - id
        - userId
        - dutySlotId
        - date
        - enteredAt
    DutyTodayResponse:
      type: object
      properties:
        slot:
          $ref: "#/components/schemas/DutySlot"
          nullable: true
        entry:
          $ref: "#/components/schemas/DutyEntry"
          nullable: true
      required:
        - slot
        - entry
    DutyEntryWithUser:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        displayName:
          type: string
          nullable: true
        dutySlotId:
          type: string
        date:
          type: string
        enteredAt:
          type: string
        confirmedAt:
          type: string
          nullable: true
      required:
        - id
        - userId
        - dutySlotId
        - date
        - enteredAt
...
    CreateCustomRoleRequest:
      type: object
      properties:
        name:
          type: string
        color:
          type: string
        permissions:
          type: array
          items:
            type: string
      required:
        - name
        - permissions
    UpdateCustomRoleRequest:
      type: object
      properties:
        name:
          type: string
        color:
          type: string
        permissions:
          type: array
          items:
            type: string
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with IServ credentials
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Account not active
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /duty:
    get:
      operationId: getDutySlots
      tags: [duty]
      summary: List duty slots
      responses:
        "200":
          description: List of duty slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DutySlot"
    post:
      operationId: createDutySlot
      tags: [duty]
      summary: Create duty slot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDutySlotRequest"
      responses:
        "201":
          description: Duty slot created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DutySlot"
  /duty/{id}:
    delete:
      operationId: deleteDutySlot
      tags: [duty]
      summary: Delete duty slot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /duty/today:
    get:
      operationId: getDutyToday
      tags: [duty]
      summary: Get current user's duty slot and entry for today
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Today's duty status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DutyTodayResponse"
  /duty/enter:
    post:
      operationId: enterDuty
      tags: [duty]
      summary: Enter duty for today
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Duty entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DutyEntry"
        "403":
          description: No duty slot today
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already entered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /duty/confirm:
    post:
      operationId: confirmDuty
      tags: [duty]
      summary: Confirm daily attendance
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Confirmation recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DutyEntry"
        "404":
          description: No entry today
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /duty/entries:
    get:
      operationId: getDutyEntries
      tags: [duty]
      summary: List all duty entries (managers only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of duty entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DutyEntryWithUser"
...
  /custom-roles:
    get:
      operationId: getCustomRoles
      tags: [roles]
      summary: List custom roles
      responses:
        "200":
          description: List of custom roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomRole"
    post:
      operationId: createCustomRole
      tags: [roles]
      summary: Create custom role (owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomRoleRequest"
      responses:
        "201":
          description: Custom role created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomRole"
  /custom-roles/{id}:
    patch:
      operationId: updateCustomRole
      tags: [roles]
      summary: Update custom role (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomRoleRequest"
      responses:
        "200":
          description: Custom role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomRole"
    delete:
      operationId: deleteCustomRole
      tags: [roles]
      summary: Delete custom role (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...